Modificaciones en cierre de caja
Se modifico el modelo factura

+++ v_ejb/ejbModule/v/eao/FacturaVentaEao.java
@@ -90,7 +90,7 @@ public class FacturaVentaEao implements FacturaVentaEaoLocal {
 	}
 	
 	@Override
-	public FacturaVenta findById(Long id) {
+	public FacturaVenta findById(Integer id) {
 		return em.find(FacturaVenta.class, id);
 	}	
 }
